+++ client-runtime/src/test/java/com/microsoft/rest/v2/policy/RequestIdPolicyTests.java
@@ -6,7 +6,6 @@
 
 package com.microsoft.rest.v2.policy;
 
-import com.microsoft.rest.v2.http.HttpClient;
 import com.microsoft.rest.v2.http.HttpHeaders;
 import com.microsoft.rest.v2.http.HttpMethod;
 import com.microsoft.rest.v2.http.HttpPipeline;
@@ -14,13 +13,9 @@ import com.microsoft.rest.v2.http.HttpPipelineBuilder;
 import com.microsoft.rest.v2.http.HttpRequest;
 import com.microsoft.rest.v2.http.HttpResponse;
 import com.microsoft.rest.v2.http.MockHttpClient;
-import com.microsoft.rest.v2.policy.RequestIdPolicy;
-import com.microsoft.rest.v2.policy.RequestPolicy;
-import com.microsoft.rest.v2.policy.RetryPolicyFactory;
 import io.reactivex.Flowable;
 import org.junit.Assert;
 import org.junit.Test;
-import io.reactivex.Observable;
 import io.reactivex.Single;
 
 import java.io.IOException;
@@ -75,7 +70,7 @@ public class RequestIdPolicyTests {
     @Test
     public void newRequestIdForEachCall() throws Exception {
         HttpPipeline pipeline = new HttpPipelineBuilder()
-            .withRequestPolicy(new RequestIdPolicy.Factory())
+            .withRequestPolicy(new RequestIdPolicyFactory())
             .withHttpClient(new MockHttpClient() {
                 String firstRequestId = null;
                 @Override
@@ -120,7 +115,7 @@ public class RequestIdPolicyTests {
                     return Single.just(mockResponse);
                 }
             },
-            new RequestIdPolicy.Factory(),
+            new RequestIdPolicyFactory(),
             new RetryPolicyFactory(1));
 
         pipeline.sendRequestAsync(new HttpRequest("sameRequestIdForRetry", HttpMethod.GET, new URL("http://localhost/"))).blockingGet();
